renamed files, moved plotting funcs to processing
