+++ ideas_modules/ideas_processing.py
@@ -6,12 +6,13 @@ import pandas as pd
 import time
 import requests
 
+import matplotlib.pyplot as plt
+
 dt_format = "%Y-%m-%dT%H:%M:%SZ"
 
 '''
 IDEAS endpoint functions
-'''
-
+''' 
 
 def spatial_timeseries(base_url: str, dataset: str, bb: dict, start_time: datetime, end_time: datetime) -> xr.Dataset:
     '''
@@ -387,3 +388,168 @@ def hofmoeller_prep(var_json: dict, dim: str) -> xr.Dataset:
         )
     )
     return ds
+
+###############################
+
+
+def spatial_mean(base_url, dataset, bb, start_time, end_time, proc=[]):
+    url = '{}/timeSeriesSpark?ds={}&minLon={}&minLat={}&maxLon={}&maxLat={}&startTime={}&endTime={}&lowPassFilter=False&seasonalFilter=False'.\
+        format(base_url, dataset, bb['min_lon'], bb['min_lat'], bb['max_lon'], bb['max_lat'],
+               start_time.strftime(dt_format), end_time.strftime(dt_format))
+
+    # Display some information about the job
+    print(url); print()
+
+    # Query SDAP to compute the time averaged map
+    print("Waiting for response from SDAP...")
+    start = time.perf_counter()
+    ts_json = requests.get(url, verify=False).json()
+    print("Time series took {} seconds".format(time.perf_counter() - start))
+    return prep_ts(ts_json, proc)
+
+def prep_ts(ts_json, proc):
+    shortname = ts_json['meta'][0]['shortName']
+    time = np.array([np.datetime64(ts[0]["iso_time"][:10]) for ts in ts_json["data"]])
+    vals = np.array([ts[0]["mean"] for ts in ts_json["data"]])
+    
+    da = xr.DataArray(vals, coords = [time], dims=['time'])
+    
+    for proc in proc:
+        da = proc(da)
+        
+    da.attrs['shortname'] = shortname
+        
+    return da
+
+def calc_anoms(data):
+    return data - np.nanmean(data)
+
+def comparison_plot(data, x_label, y_label, var='', anoms=False):
+    plt.figure(figsize=(15,6))
+    
+    for da in data:
+        if anoms:
+            vals = calc_anoms(da.values)
+        else:
+            vals = da.values
+        plt.plot(da.time, vals, linewidth=2, label=da.attrs['shortname'])
+    
+    plt.grid(b=True, which='major', color='k', linestyle='-')
+    plt.xlabel(x_label, fontsize=12)
+    plt.ylabel (y_label, fontsize=12)
+    plt.xticks(rotation=45)
+    plt.title(f'{var}{" Anomalies" if anoms else ""}', fontsize=16)
+    plt.legend(prop={'size': 12})
+    plt.show()
+    
+def temporal_variance(base_url, dataset, bb, start_time, end_time):
+    params = {
+        'ds': dataset,
+        'minLon': bb['min_lon'],
+        'minLat': bb['min_lat'],
+        'maxLon': bb['max_lon'],
+        'maxLat': bb['max_lat'],
+        'startTime': start_time.strftime(dt_format),
+        'endTime': end_time.strftime(dt_format)
+    }
+    
+    url = '{}/varianceSpark?ds={}&minLon={}&minLat={}&maxLon={}&maxLat={}&startTime={}&endTime={}'.\
+        format(base_url, dataset, bb['min_lon'], bb['min_lat'], bb['max_lon'], bb['max_lat'],
+               start_time.strftime(dt_format), end_time.strftime(dt_format))
+    
+    # Display some information about the job
+    print(url); print()
+    
+    # Query SDAP to compute the time averaged map
+    print("Waiting for response from SDAP...")
+    start = time.perf_counter()
+    var_json = requests.get(url, params=params, verify=False).json()
+    print("Time series took {} seconds".format(time.perf_counter() - start))
+    return prep_var(var_json)
+    
+def prep_var(var_json):
+    shortname = var_json['meta']['shortName']
+
+    vals = np.array([v['variance'] for var in var_json['data'] for v in var])
+    lats = np.array([var[0]['lat'] for var in var_json['data']])
+    lons = np.array([v['lon'] for v in var_json['data'][0]])
+    
+    vals[vals==-9999]=np.nan
+    
+    vals_2d = np.reshape(vals, (len(var_json['data']), len(var_json['data'][0])))
+
+    da = xr.DataArray(vals_2d, coords={"lat": lats, "lon": lons}, dims=["lat", "lon"])
+    da.attrs['shortname'] = shortname
+    da.attrs['units'] = '$m^2/s^2$'
+    return da
+
+def get_in_situ_data(start_time: str, end_time: str,
+                     min_lon: int, max_lon: int, min_lat: int, max_lat: int, provider: str) -> pd.DataFrame:
+    data = []
+    query_url = f'{url_in_situ}/{endpoint_in_situ}?' \
+                f'startIndex={start_index_in_situ}&itemsPerPage={items_per_page_in_situ}&' \
+                f'startTime={start_time}&endTime={end_time}&' \
+                f'bbox={min_lon},{min_lat},{max_lon},{max_lat}&' \
+                f'provider={provider}'
+    if query_url:
+        print(query_url)
+    while query_url:
+        resp = requests.get(query_url, verify=False).json()
+        if len(resp['results']):
+            data += resp['results']
+        query_url = resp['next'].replace('http:', 'https:') if resp['last'] != resp['next'] else None
+
+    
+    return pd.DataFrame(data) if len(data) else None
+
+def stacked_overlay_plot(x_datas: List[np.array], y_datas: List[np.array],
+                series_labels: List[str], y_labels=List[str], title: str='',
+                top_paddings: List[int]=[0, 0]):
+
+    plt.style.use('ggplot')
+    fig, ax = plt.subplots(2, 1, sharex=True)
+
+    # Plot 1
+    ax[0].set_title(title)
+    ax[0].plot(
+        [ datetime.strptime(x_val, '%Y-%m-%dT%H:%M:%SZ').replace(year=2022) for x_val in x_datas[0] ],
+        y_datas[0], label=series_labels[0])
+        
+    # Plot 2
+    ax[0].plot(
+        [ datetime.strptime(x_val, '%Y-%m-%dT%H:%M:%SZ').replace(year=2022) for x_val in x_datas[1] ],
+        y_datas[1], label=series_labels[1])
+
+    ax[0].legend(loc='upper center', shadow=True)
+    y_data_max = max( np.amax(y_datas[0]), np.amax(y_datas[1]) )
+    ax[0].set_ylim([ 0, y_data_max + top_paddings[0] ])
+    ax[0].set_ylabel(y_labels[0])
+
+    # Plot 3
+    ax[1].plot(
+        [ datetime.strptime(x_val, '%Y-%m-%dT%H:%M:%SZ').replace(year=2022) for x_val in x_datas[2] ],
+        y_datas[2], label=series_labels[2])
+        
+    # Plot 4
+    ax[1].plot(
+        [ datetime.strptime(x_val, '%Y-%m-%dT%H:%M:%SZ').replace(year=2022) for x_val in x_datas[3] ],
+        y_datas[3], label=series_labels[3])
+    
+    ax[1].legend(loc='upper center', shadow=True)
+    y_data_max = max(np.amax(y_datas[2]), np.amax(y_datas[3]))
+    ax[1].set_ylim([ 0, y_data_max + top_paddings[1] ])
+    ax[1].set_ylabel(y_labels[1])
+
+    # Set title and legend
+    plt.legend(loc='upper center', shadow=True)
+
+    # Set grid and ticks
+    dtFmt = mdates.DateFormatter('%b %d')
+    plt.gca().xaxis.set_major_formatter(dtFmt)
+    plt.xticks(rotation=45)
+    ax[0].tick_params(left=False, bottom=False)
+    ax[1].tick_params(left=False, bottom=False)
+    ax[0].grid(b=True, which='major', color='k', linestyle='--', linewidth=0.25)
+    ax[1].grid(b=True, which='major', color='k', linestyle='--', linewidth=0.25)
+    
+    plt.show()
